requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
youtube-dl>=2021.12.17
PyNaCl>=1.5.0
yt-dlp>=2023.7.6
asyncio>=3.4.3